requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
cramjam==2.5.0
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.12.0
django-environ==0.9.0
django-redis==5.2.0
djangorestframework==3.13.1
fastparquet==0.8.3
filelock==3.7.1
fsspec==2022.8.2
gunicorn==20.1.0
idna==3.4
jsonfield==3.1.0
mergedeep==1.3.4
numpy==1.22.4
packaging==21.3
pandas==1.4.2
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyarrow==8.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.15.1
whitenoise==6.1.0
wrapt==1.14.1
